pyproject.toml
[project]
name = "sprint-marketing-bot"
version = "0.1.0"
description = "Local RAG chatbot with strict provenance and zero cloud runtime cost"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.0",
  "sentence-transformers>=3.0.1",
  "faiss-cpu>=1.8.0",
  "numpy>=1.26.0",
  "pandas>=2.2.0",
  "pypdf>=4.2.0",
  "python-docx>=1.1.0",
  "markdown>=3.6",
  "beautifulsoup4>=4.12.3",
  "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.5.0",
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.7",
]

[tool.ruff]
line-length = 100

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]


